requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
google-adk==1.28.0
google-genai==1.69.0
python-dotenv==1.2.2